requirements.txt
# Web Server
fastapi
uvicorn
python-multipart
jinja2

# Environment and Utils
python-dotenv
requests
pydantic
nest_asyncio

# AI and ML
google-genai
# genai==0.3.0
google-generativeai
sentence-transformers
llama-index-core
llama-index-embeddings-huggingface
llama-parse
chromadb


# Google Cloud Services
google-cloud-texttospeech

# Additional dependencies
aiohttp

# WebSocket
websockets==14.1
asyncio==3.4.3

# Gradio
gradio-client


